move decorations ordering method to utils
